chore(mobile): clean up SIP9 model and only model data we use
