+++ packages/models/src/assets/asset.model.ts
@@ -1,4 +1,5 @@
 import { InscriptionMimeType } from '../inscription-mime-type.model';
+import { Sip9Asset } from './sip9-asset.model';
 
 export const CryptoAssetChains = {
   bitcoin: 'bitcoin',
@@ -102,7 +103,7 @@ export type FungibleCryptoAsset =
   | RuneAsset;
 
 // NFT asset types
-interface BaseNonFungibleCryptoAsset extends BaseCryptoAsset {
+export interface BaseNonFungibleCryptoAsset extends BaseCryptoAsset {
   readonly category: 'nft';
   readonly protocol: NonFungibleCryptoAssetProtocol;
 }
@@ -131,151 +132,6 @@ export interface StampAsset extends BaseNonFungibleCryptoAsset {
   readonly stampUrl: string;
 }
 
-export type Sip9ContentType =
-  // Images
-  | 'image/jpeg'
-  | 'image/jpg'
-  | 'image/png'
-  | 'image/gif'
-  | 'image/webp'
-  | 'image/svg+xml'
-  | 'image/bmp'
-  | 'image/tiff'
-  | 'image/avif'
-
-  // Videos
-  | 'video/mp4'
-  | 'video/webm'
-  | 'video/mov'
-  | 'video/quicktime'
-  | 'video/avi'
-  | 'video/x-msvideo'
-  | 'video/ogg'
-
-  // Audio
-  | 'audio/mpeg'
-  | 'audio/mp3'
-  | 'audio/wav'
-  | 'audio/x-wav'
-  | 'audio/ogg'
-  | 'audio/aac'
-  | 'audio/flac'
-  | 'audio/webm'
-
-  // 3D Models
-  | 'model/gltf+json'
-  | 'model/gltf-binary'
-  | 'application/octet-stream' // GLB files sometimes use this
-
-  // Documents/Text
-  | 'text/plain'
-  | 'text/html'
-  | 'text/markdown'
-  | 'application/pdf'
-  | 'application/json'
-
-  // Interactive/Web
-  | 'text/javascript'
-  | 'application/javascript'
-
-  // Archives
-  | 'application/zip'
-
-  // Unknown/Fallback
-  | 'unknown';
-
-export type SupportedSip9ContentType =
-  | Extract<
-      Sip9ContentType,
-      | 'image/jpeg'
-      | 'image/png'
-      | 'image/gif'
-      | 'image/webp'
-      | 'image/svg+xml'
-      | 'image/bmp'
-      | 'image/tiff'
-      | 'image/avif'
-      | 'video/mp4'
-      | 'audio/mpeg'
-      | 'audio/mp3'
-      | 'audio/wav'
-      | 'audio/ogg'
-      | 'audio/aac'
-      | 'audio/flac'
-      | 'audio/webm'
-      | 'text/plain'
-      | 'application/octet-stream'
-      | 'model/gltf+json'
-      | 'model/gltf-binary'
-    >
-  | '';
-
-export interface Sip9Collection {
-  id: string;
-  type?: string;
-  name: string;
-  isVerified: boolean;
-  locationUrl: string;
-  totalItems?: number;
-  floorPrice?: {
-    amount: number;
-    unit: string;
-  };
-}
-
-export interface Sip9Owner {
-  address: string;
-  chain: string;
-  id: string;
-  displayName: string;
-  slug: string;
-  avatarUrl: string | null;
-  avatarContentType: string | null;
-  profileUrl: string;
-  bio: string | null;
-  isVerified: boolean;
-}
-
-export interface Sip9AssetContent {
-  contentUrl: string;
-  contentType: string;
-}
-
-export interface Sip9Attribute {
-  traitType: string;
-  displayType?: string;
-  value: any;
-  rarityPercent?: number;
-}
-
-export interface Sip9Details {
-  id?: string;
-  name: string;
-  description: string;
-  assetContent?: Sip9AssetContent;
-  cachedImage: string;
-  cachedImageThumbnail: string;
-  contentType: SupportedSip9ContentType;
-  locationUrl?: string;
-  collection?: Sip9Collection;
-  owner?: Sip9Owner;
-  attributes?: Sip9Attribute[];
-  rarityRank?: number;
-}
-export interface Sip9Asset extends BaseNonFungibleCryptoAsset {
-  readonly chain: 'stacks';
-  readonly protocol: 'sip9';
-  readonly assetId: string;
-  readonly contractId: string;
-  readonly tokenId: number;
-  readonly name: string;
-  readonly description: string;
-  readonly cachedImage: string;
-  readonly cachedImageThumbnail: string;
-  readonly contentType: SupportedSip9ContentType;
-  readonly details: Sip9Details;
-  readonly collection?: Sip9Collection;
-}
 export type NonFungibleCryptoAsset = InscriptionAsset | StampAsset | Sip9Asset;
 
 export type CryptoAsset = FungibleCryptoAsset | NonFungibleCryptoAsset;
