+++ packages/services/src/assets/sip9-asset.service.ts
@@ -4,11 +4,8 @@ import { Sip9Asset } from '@leather.io/models';
 
 import { GammaApiClient } from '../infrastructure/api/gamma/gamma-api.client';
 import { HiroStacksApiClient } from '../infrastructure/api/hiro/hiro-stacks-api.client';
-import {
-  createSip9Asset,
-  getContractPrincipalFromAssetIdentifier,
-  getNonFungibleTokenId,
-} from './stacks-asset.utils';
+import { getContractPrincipalFromAssetIdentifier } from './stacks-asset.utils';
+import { createSip9Asset, getNonFungibleTokenId } from './sip9-asset.utils';
 
 @injectable()
 export class Sip9AssetService {
